Restructure loan feature tests
